разнесены модели по файлам, не используемые поля модели удалены
